temp_requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.1.0
camelot-py==1.0.9
certifi==2025.8.3
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
cryptography==45.0.6
distro==1.9.0
et_xmlfile==2.0.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
jpype1==1.6.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
langcodes==3.5.0
language_data==1.3.0
locate==1.1.1
marisa-trie==1.3.0
MarkupSafe==3.0.2
msgpack==1.1.1
narwhals==2.1.2
numpy==2.2.6
openai==1.101.0
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.2
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
pip==25.1.1
plotly==6.3.0
protobuf==6.32.0
pyarrow==21.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
PyMuPDF==1.26.3
pypdf==5.9.0
PyPDF2==3.0.1
pypdfium2==4.30.0
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2025.7.34
requests==2.32.5
rpds-py==0.27.0
selenium==4.35.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
streamlit==1.48.1
tabula-py==2.10.0
tabulate==0.9.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
wcwidth==0.2.13
websocket-client==1.8.0
wordfreq==3.1.1
wsproto==1.2.0
xlsxwriter==3.2.5
